A little cleanup
