+++ ConsoleMarkdownRenderer.Example/ConsoleMarkdownRenderer.Example.csproj
@@ -8,6 +8,7 @@
     <OutputType>Exe</OutputType>
     <TargetFrameworks>net8.0;net10.0</TargetFrameworks>
     <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
   <ItemGroup>
