+++ ConsoleMarkdownRenderer.Example/Program.cs
@@ -1,9 +1,5 @@
-﻿using System;
-using System.ComponentModel;
+﻿using System.ComponentModel;
 using System.Diagnostics.CodeAnalysis;
-using System.IO;
-using System.Threading;
-using System.Threading.Tasks;
 using Spectre.Console;
 using Spectre.Console.Cli;
