+++ ConsoleMarkdownRenderer.Fakes/ConsoleMarkdownRenderer.Fakes.csproj
@@ -7,6 +7,7 @@
     <EmbedUntrackedSources>true</EmbedUntrackedSources>
     <DebugType>embedded</DebugType>
     <Deterministic>true</Deterministic>
+    <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
   <PropertyGroup>
