+++ ConsoleMarkdownRenderer.Tests/DownLoadTests.cs
@@ -1,10 +1,5 @@
-using System;
-using System.IO;
 using System.Net;
-using System.Net.Http;
 using System.Text;
-using System.Threading.Tasks;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace ConsoleMarkdownRenderer.Tests
 {
