+++ ConsoleMarkdownRenderer.Tests/FakeHttpHelpers.cs
@@ -1,8 +1,3 @@
-using System;
-using System.Net.Http;
-using System.Threading;
-using System.Threading.Tasks;
-
 namespace ConsoleMarkdownRenderer.Tests
 {
     /// <summary>
