+++ ConsoleMarkdownRenderer.Tests/HandleLinkItemTests.cs
@@ -1,7 +1,3 @@
-using System;
-using System.IO;
-using System.Threading.Tasks;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Microsoft.VisualStudio.TestTools.UnitTesting.Logging;
 
 namespace ConsoleMarkdownRenderer.Tests
