+++ ConsoleMarkdownRenderer.Tests/PromptResultTests.cs
@@ -1,6 +1,3 @@
-using System;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
 namespace ConsoleMarkdownRenderer.Tests
 {
     /// <summary>
