+++ ConsoleMarkdownRenderer.Tests/StylingTests.cs
@@ -1,7 +1,5 @@
-using System;
 using System.Reflection;
 using ConsoleMarkdownRenderer.Styling;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Spectre.Console;
 
 namespace ConsoleMarkdownRenderer.Tests
