+++ ConsoleMarkdownRenderer.Tests/TempFileManagerTests.cs
@@ -1,6 +1,3 @@
-using System.IO;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
 namespace ConsoleMarkdownRenderer.Tests
 {
     /// <summary>
