+++ IMarkdownDisplayer.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Threading.Tasks;
-
 namespace ConsoleMarkdownRenderer
 {
     /// <summary>
