+++ MarkdownDisplayer.cs
@@ -1,14 +1,7 @@
 using System.Runtime.CompilerServices;
-using System;
-using System.Collections.Generic;
 using System.Diagnostics;
-using System.IO;
-using System.Linq;
 using System.Net;
-using System.Net.Http;
 using System.Runtime.InteropServices;
-using System.Threading;
-using System.Threading.Tasks;
 using ConsoleMarkdownRenderer.ObjectRenderers;
 using Markdig;
 using Spectre.Console;
