+++ ObjectRenderers/ConsoleRendererBase.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Text;
 using Markdig.Helpers;
 using Markdig.Renderers;
