+++ ObjectRenderers/TempFileManager.cs
@@ -1,7 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.IO;
-
 namespace ConsoleMarkdownRenderer.ObjectRenderers
 {
     /// <summary>
