+++ ObjectRenderers/Utilities.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Collections.Generic;
-
 namespace ConsoleMarkdownRenderer.ObjectRenderers
 {
     public static class Utilities
