+++ PromptResult.cs
@@ -1,4 +1,3 @@
-using System;
 using Spectre.Console;
 
 namespace ConsoleMarkdownRenderer
