+++ Styling/TextColor.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace ConsoleMarkdownRenderer.Styling
 {
     /// <summary>
