+++ Styling/TextStyle.cs
@@ -1,7 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-
 namespace ConsoleMarkdownRenderer.Styling
 {
     /// <summary>
