+++ Styling/TextStyleExtensions.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using Spectre.Console;
 
 namespace ConsoleMarkdownRenderer.Styling
